feat: add Populate method for in-place deserialization with comprehensive error handling and validation
